Updated screen output and source files to reflect version 2.0.

+++ garli/branches/partitioning/src/errorexception.h
@@ -34,7 +34,7 @@ class ErrorException{
 	
 	public:
 	char *message;
-	int messlen;
+	size_t messlen;
 	//char message[5000];
 	//char message[400];
 	ErrorException(){
